Create TBC Scroll Origins NFT Contract
NOTE: This is just a web3 smart contract. It's not an actual contract but a test project to interact on the Scroll mainnet launch.

After launching three successful testnets on Scroll Alpha Testnet, Scroll Mainnet is ready for developers and protocols to deploy their contracts and explore Scroll’s newly launched Mainnet. To commemorate this moment, Scroll is launching Scroll Origins, an NFT mint to celebrate Scroll’s earliest builders. This makes a permanent on-chain record that you were here with us at this critical moment.

DISCLAIMER: Use at your own risk and NFA!